requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
decorator==5.2.1
distro==1.9.0
exceptiongroup==1.3.0
executing==2.2.0
frozenlist==1.7.0
genai==2.1.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
ipython==8.18.1
jedi==0.19.2
jiter==0.10.0
matplotlib-inline==0.1.7
multidict==6.6.4
openai==0.27.10
parso==0.8.5
pexpect==4.9.0
prompt_toolkit==3.0.52
propcache==0.3.2
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
regex==2025.7.34
requests==2.32.5
sniffio==1.3.1
stack-data==0.6.3
tabulate==0.9.0
tiktoken==0.3.3
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
wcwidth==0.2.13
yarl==1.20.1
